+++ kse/src/main/java/org/kse/gui/actions/ExportKeyPairPublicKeyAction.java
@@ -34,7 +34,7 @@ import javax.swing.ImageIcon;
 import javax.swing.JOptionPane;
 
 import org.kse.crypto.CryptoException;
-import org.kse.crypto.publickey.JwkPublicKey;
+import org.kse.crypto.publickey.JwkPubExporter;
 import org.kse.crypto.publickey.OpenSslPubUtil;
 import org.kse.crypto.x509.X509CertUtil;
 import org.kse.gui.KseFrame;
@@ -95,7 +95,7 @@ public class ExportKeyPairPublicKeyAction extends KeyStoreExplorerAction {
                     encoded = OpenSslPubUtil.get(publicKey);
                     break;
                 case JWK:
-                    encoded = JwkPublicKey.from(publicKey, alias).get();
+                    encoded = JwkPubExporter.from(publicKey, alias).get();
                     break;
             }
